pyproject.toml
[project]
name = "air-pollution"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jupyter>=1.1.1",
    "python-dotenv>=1.1.0",
    "httpx>=0.28.1",
    "numpy>=2.2.4",
    "pandas>=2.2.3",
    "plotly>=6.0.1",
    "tqdm>=4.67.1",
    "psycopg>=3.2.6",
    "psycopg-binary>=3.2.6",
    "psycopg-pool>=3.2.6",
    "tenacity>=9.0.0",
]
